Move schema to top level
This fixes the server crashing in Docker, since I fucked up the COPY
command and it actually copied the contents instead of the directory
itself. I wanted to avoid an extra Docker layer to keep the image
smaller and opted for moving the schema to top level instead of adding
another COPY command.